Rename the AppBar store
